working on output

+++ include/EdgeTransaction.h
@@ -9,10 +9,10 @@ typedef struct EdgeTransaction
     struct EdgeTransaction* nextIn;
     struct EdgeTransaction* nextOut;
     char* date;
-    double amount;
+    int amount;
 } EdgeTransaction;
 
-EdgeTransaction* edgeTransactionCreate(double amount, char* date,struct NodeAccount* owner,
+EdgeTransaction* edgeTransactionCreate(int amount, char* date,struct NodeAccount* owner,
  struct NodeAccount* destination);
 void edgeTransactionFree(EdgeTransaction* edgeTransaction);
 void edgeTransactionPrint(EdgeTransaction* edgeTransaction);
