+++ include/NodeAccount.h
@@ -39,8 +39,8 @@ void nodeAccountRemoveInEdge(NodeAccount* nsode, struct EdgeTransaction* edgeTra
 
 // modify edges
 
-void nodeAccountFindAndModifyEdgeWithNode(NodeAccount* fromNode, NodeAccount* toNode, double oldSum,
-double newSum, char* oldDate, char* newDate);
+void nodeAccountFindAndModifyEdgeWithNode(NodeAccount* fromNode, NodeAccount* toNode, int oldSum,
+int newSum, char* oldDate, char* newDate);
 
 // getters
